Adding tables for bookmarks and folders (#727)
* Adding tables for bookmarks and folders

* Removed deprecated code

* Added model for bookmarks, made db rerunnable

---------

+++ db/migrate/20200907000752_drop_reviews.rb
@@ -1,5 +1,7 @@
-class DropReviews < ActiveRecord::Migration[5.2]
+class DropReviews < ActiveRecord::Migration[6.1]
   def change
-    drop_table :reviews
+    if table_exists?(:reviews)
+      drop_table :reviews
+    end  
   end
 end
